requirements.txt
websockets>=10.0
python-dotenv>=0.19.0
